ocr-service/requirements.txt
paddlepaddle==2.6.1
paddleocr==2.7.3
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pillow==10.2.0
numpy==1.26.3
